# openapi.yaml
openapi: 3.1.0
info:
  title: WDC Backend
  version: 0.1.0
  description: |
    Kaduna State WDC Digital Reporting Platform — REST API.

    M1 milestone: only health endpoints are exposed. Subsequent milestones
    add auth, users, wards, forms, reports, sync, messaging, investigations,
    audit, ai, and notifications under `/api/v1/`.
servers:
  - url: http://localhost:3000
    description: Local development
paths:
  /api/v1/auth/sign-in/mobile:
    post:
      summary: Mobile sign-in (phone + PIN)
      operationId: authSignInMobile
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, pin, deviceId]
              properties:
                phone: { type: string, example: '+2348012345678' }
                pin: { type: string, pattern: '^\d{6}$' }
                deviceId: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '200': { description: token pair, content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } } }
        '401': { description: invalid credentials }
        '422': { description: validation error }
  /api/v1/auth/sign-in/console:
    post:
      summary: Console sign-in (email + password + TOTP, director only)
      operationId: authSignInConsole
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, totp, deviceId]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 12, maxLength: 256 }
                totp: { type: string, pattern: '^\d{6}$' }
                deviceId: { type: string, minLength: 1, maxLength: 64 }
      responses:
        '200': { description: token pair, content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } } }
        '401': { description: invalid credentials }
        '422': { description: validation error }
  /api/v1/auth/set-credentials:
    post:
      summary: Redeem an enrolment token to set PIN or password+TOTP
      operationId: authSetCredentials
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrolmentToken]
              properties:
                enrolmentToken: { type: string, minLength: 40, maxLength: 100 }
                pin: { type: string, pattern: '^\d{6}$', description: 'For mobile (secretary/coordinator)' }
                password: { type: string, minLength: 12, maxLength: 256, description: 'For console (director)' }
                totpSecret: { type: string, minLength: 16, maxLength: 64 }
                totp: { type: string, pattern: '^\d{6}$' }
      responses:
        '200':
          description: credentials set
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties: { ok: { type: boolean, enum: [true] } }
        '400': { description: missing required fields for the user's role }
        '401': { description: token unrecognised or expired }
  /api/v1/auth/refresh:
    post:
      summary: Rotate refresh token; returns a fresh access+refresh pair
      operationId: authRefresh
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken, deviceId]
              properties:
                refreshToken: { type: string, minLength: 32, maxLength: 256 }
                deviceId: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '200': { description: new token pair, content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } } }
        '401': { description: token unrecognised, expired, mismatched device, or reused }
  /api/v1/auth/sign-out:
    post:
      summary: Revoke all refresh tokens for the calling device
      operationId: authSignOut
      tags: [auth]
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceId]
              properties:
                deviceId: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '200':
          description: count of refresh tokens revoked
          content:
            application/json:
              schema:
                type: object
                required: [revoked]
                properties:
                  revoked: { type: integer, minimum: 0 }
        '401': { description: missing or invalid bearer }
...
  /api/v1/audit/anchor:
    post:
      summary: Create a new audit anchor (director / scheduler)
      operationId: auditCreateAnchor
      tags: [audit]
      security: [{ bearer: [] }]
      responses:
        '200':
          description: anchor created (or no-op when audit log is empty)
          content:
            application/json:
              schema:
                type: object
                required: [created]
                properties:
                  created: { type: boolean }
                  id: { type: string }
                  latestEventId: { type: string }
                  latestHash: { type: string }
        '403': { description: requires director / system }
  /api/v1/audit/anchors:
    get:
      summary: List recent audit anchors with verify-on-read status
      operationId: auditListAnchors
      tags: [audit]
      security: [{ bearer: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        '200':
          description: anchors (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [id, anchoredAt, latestEventId, latestHash, signatureAlg, signingKeyId, signature, verified]
                  properties:
                    id: { type: string }
                    anchoredAt: { type: string, format: date-time }
                    latestEventId: { type: string }
                    latestHash: { type: string }
                    signatureAlg: { type: string }
                    signingKeyId: { type: string }
                    signature: { type: string }
                    verified: { type: boolean }
        '403': { description: requires director / system }
  /api/v1/audit/export:
    get:
      summary: Sealed CSV export bounded by the most recent anchor
      operationId: auditExport
      tags: [audit]
      security: [{ bearer: [] }]
      responses:
        '200':
          description: CSV stream prefixed with anchor preamble (#-prefixed lines)
          content:
            text/csv:
              schema: { type: string }
        '403': { description: requires director / system }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TokenResponse:
      type: object
      required: [accessToken, accessExpiresIn, refreshToken, refreshExpiresAt]
      properties:
        accessToken: { type: string, description: 'JWT (RS256), 15min default TTL' }
        accessExpiresIn: { type: integer, description: 'Seconds until access expires' }
        refreshToken: { type: string, description: 'Opaque token; rotated on use' }
        refreshExpiresAt: { type: string, format: date-time }
...